% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoPath.R
\name{drive_i}
\alias{drive_i}
\title{drive_i}
\usage{
drive_i(s)
}
\arguments{
\item{s}{numeric}
}
\value{
numeric
}
\description{
Costs for Wheeled Vehicles
}
\details{
The `drive_i` function will calculate a cost surface, where costs refer to the use of
wheeled vehicles with a critical slope value of 8%. Here the 
adapted function published by Herzog 2012 baesd on Llobera 2007 is used.

Note: As gdistance uses resistor instead of cost, the function creates inverse costs
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>

Hendrik Raese <\email{h.raese@ufg.uni-kiel.de}>
}
