% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{edist}
\alias{edist}
\title{edist}
\usage{
edist(x)
}
\arguments{
\item{x}{numeric}
}
\description{
Euclidian Distance in multidimensional space
}
\details{
Calculates the euclidian distance in a multidimensional space
}
\examples{
 
 x <- c(2,4,1,5,7,8)
 edist(x)

}
\author{
Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
