% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{factor}
\alias{factor}
\title{factor}
\usage{
factor(x, a, b)
}
\arguments{
\item{x}{Variable}

\item{a}{Variable}

\item{b}{Variable}
}
\description{
Linear function to scale density values
}
\details{
Description
}
\author{
Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
