% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{gau1}
\alias{gau1}
\title{gau1}
\usage{
gau1(x, sd)
}
\arguments{
\item{x}{numeric, vector}

\item{sd}{numeric, standard deviation}
}
\value{
density value of normal distribution
}
\description{
Factor defining size of first Kernel
}
\details{
factor defining size of the first kernel, which generate the stucture of dynamic kernel
}
\examples{
 
 x <- c(2,4,1,5,7,8)
 sd <- mean(x)
 gau1(x, sd)

}
\author{
Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
