% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoPath.R
\name{hdiff}
\alias{hdiff}
\title{hdiff}
\usage{
hdiff(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric
}
\description{
Calculates Slope Values out of Elevation
}
\details{
The `hdiff` function will calculate slope based on the numeric value of the elevation
in adjacent cells
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
