% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{kernel.par}
\alias{kernel.par}
\title{kernel.par}
\usage{
kernel.par(xp, s, int)
}
\arguments{
\item{xp}{numeric, scaled density values multiplied by nearest neighbourhood distance}

\item{s}{numeric, Kernelparameter: incline starting from point 1}

\item{int}{numeric, scaled density values}
}
\description{
kernel.par
}
\details{
Description
}
\examples{
 
 xp <- 6
 s <- -0.3
 int <- 7
 kernel.par(xp,s,int)

}
\author{
Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
