% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{kernel1d}
\alias{kernel1d}
\title{kernel1d}
\usage{
kernel1d(x, kp)
}
\arguments{
\item{x}{Variable}

\item{kp}{Variable}
}
\description{
KDE Function
}
\details{
Description
}
\author{
Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
