% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{makestatkde}
\alias{makestatkde}
\title{makestatkde}
\usage{
makestatkde(pppm, f_sd1 = 4, sgdf, df, x = 1, y = 2, num)
}
\arguments{
\item{pppm}{PointPattern object of Spatial Data}

\item{f_sd1}{factor defining size of the first kernel, which generate the stucture of dynamic kernel}

\item{sgdf}{grid object to store results}

\item{df}{data frame, containing coordinates of monuments}

\item{x}{numeric, indicating column number of x coordinates}

\item{y}{numeric, indicating column number of y coordinates}
}
\value{
SpatialGrid Object
}
\description{
Calculating Static KDE
}
\details{
description of function
}
\examples{
# Crating Test Data Randomly
testmatrix <- data.frame(x = abs(rnorm(100)*50), y = abs(rnorm(100)*50))
# Setting geographical frame
xmin    <- 0
xmax    <- max(testmatrix$x)
ymin    <- 0
ymax    <- max(testmatrix$y)
ext_ai <- extent(xmin, xmax, ymin, ymax)

sv <- (xmax-xmin)/(ymax-ymin)
rw     <- 10   # width of raster defined in m

rows  <- round((ymax-ymin)/rw, 0) + 1                                    
colums <- round((xmax-xmin)/rw, 0) + 1                                      
v <- cbind(1:(colums*rows))                                              
df <- data.frame(v)                                                         
gt      <- sp::GridTopology(c(xmin, ymin), c(rw, rw), c(colums, rows))
sgdf    <- sp::SpatialGridDataFrame(gt, df)

pppm <- spatstat::ppp(testmatrix[,1], testmatrix[,2], 
                       window = spatstat::owin(
                       xrange=c(sgdf@bbox[1,1],sgdf@bbox[1,2]),
                       yrange=c(sgdf@bbox[2,1],sgdf@bbox[2,2]), 
                       unitname="m"))
                       
base_kde <- makestatkde(pppm, sgdf=sgdf, df=testmatrix, x=1, y=2, num=length(df[,1]))

}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
