% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoPath.R
\name{theo_del}
\alias{theo_del}
\title{theo_del}
\usage{
theo_del(maxima, win)
}
\arguments{
\item{maxima}{SpatialPoints object}

\item{win}{owin object}
}
\value{
list with nodes and edges of centers of infrastructure
}
\description{
Calculating a Delaunay triangulation and preparing coordinates for rWalk loops
}
\details{
In absence of known contemporary nodes, like settlements, denser monument
locations can be used to reconstruct nodes of infrastructure. This function
will use the Kernel Density Approach combined with a moving window, in 
order to detect nodes in areas with less monuments (from a global perspective).
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
