% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoPath.R
\name{theoPath_herzog}
\alias{theoPath_herzog}
\title{theoPath_herzog}
\usage{
theoPath_herzog(ras_ai, con, method, theta, p, type = "c")
}
\arguments{
\item{ras_ai}{RasterLayer, raster with elevation values}

\item{con}{data.frame, connections of a Delaunay triangulation as a result of function theo_del or your own method}

\item{method}{chr, either "walk_i" for pedestrians or "drive_i" for vehicles. For further informations look up the respective functions}

\item{theta}{numeric, parameter controls randomisation of walk. Lower values equal more exploration of the walker around the shortest path, while if theta approaches zero the walk becomes totally random}

\item{p}{numeric, buffer zone around rWalk rasters, used in loop}

\item{type}{chr, either "c" (default) for least-cost distances or "r" for random walks. As stated by J. van Etten, there is no analytical way as of now to decide for intermediate values of theta which type should be choosen. For further informations see ?gdistance::geoCorrection}
}
\value{
raster object, with values of summed up expectations of single rWalk connections
}
\description{
Theoretical Paths using cost functions defined by Herzog 2012
}
\details{
If there are no actual parts of a path network known, the `theoPath_herzog` function can be used to reconstruct pathways based on randomised shortest paths connecting known regions with higher densities of sites, e.g. monuments. An underlying cost surface is created by using the cost functions defined by I. Herzog (2012) either for walking or driving. This function is a useful step in the evaluation of reconstructed paths.
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>

Hendrik Raese <\email{h.raese@roots.uni-kiel.de}>
}
