% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoPath.R
\name{walk_i}
\alias{walk_i}
\title{walk_i}
\usage{
walk_i(s)
}
\arguments{
\item{s}{numeric}
}
\value{
numeric
}
\description{
Costs for Energy Expenditure for Pedestrians
}
\details{
The `walk_i` function will calculate a cost surface, where costs refer to energy
expenditure for a pedestrian. Initially published by Llobera & ... Here the 
adapted function published by Herzog 2012 is used.

Note: As gdistance uses resistor instead of cost, the function creates inverse costs
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
