% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoPath.R
\name{localMax}
\alias{localMax}
\title{localMax}
\usage{
localMax(df, x = 1, y = 2, r = 5000, sw = 10)
}
\arguments{
\item{df}{data frame, containing coordinates of path associated features}

\item{x}{numeric, indicating column number of x coordinates}

\item{y}{numeric, indicating column number of y coordinates}

\item{r}{numeric, size of moving window in meter, default = 5000}
}
\value{
SpatialPoints
}
\description{
Local Centers of Infrastructure from Monument Location
}
\details{
In absence of known contemporary nodes, like settlements, the density of monument
locations can be used to reconstruct nodes of infrastructre. This function
will use the Kernel Density Approach combined with a moving window, in 
order to detect nodes in areas with less monuments (from a global perspective).
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
