% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repath.R
\name{repath}
\alias{repath}
\title{repath}
\usage{
repath(df, sgdf, x = 1, y = 2, colums)
}
\arguments{
\item{df}{data frame, containing coordinates of path associated features}

\item{sgdf}{SpatialGridDataFrame, same extent then area of interest}

\item{x}{numeric, indicating column number of x coordinates}

\item{y}{numeric, indicating column number of y coordinates}

\item{colums}{Muss noch in Funktion!}
}
\value{
list, first object "KDE" is a raster object of Kernel Density Estimation, 
              second object "ras_dens_ridges" coordinates of density ridges
}
\description{
Reconstructing Empirical Pathes
}
\details{
The `repath` function will calculate density ridges out of the location from 
path associated features.
}
\examples{


}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>
}
