% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoPath.R
\name{theoPath_param}
\alias{theoPath_param}
\title{theoPath_param}
\usage{
theoPath_param(ras_ai, ras_para, con, method, theta, p, type = "c")
}
\arguments{
\item{ras_ai}{RasterLayer, raster with elevation values}

\item{ras_para}{RasterLayer, raster with additional values (visibility, friction costs, etc.)}

\item{con}{data.frame, connections of a Delaunay triangulation as a result of function theo_del or your own method}

\item{method}{chr, either "walk_i" for pedestrians or "drive_i" for vehicles. For further informations look up the respective functions}

\item{theta}{numeric, parameter controls randomisation of walk. Lower values equal more exploration of the walker around the shortest path, while if theta approaches zero the walk becomes totally random}

\item{p}{numeric, buffer zone around rWalk rasters, used in loop}

\item{type}{chr, either "c" (default) for least-cost distances or "r" for random walks. As stated by J. van Etten, there is no analytical way as of now to decide for intermediate values of theta which type should be choosed. For further informations see ?gdistance::geoCorrection}
}
\value{
List, two RasterLayer with values of summed up expectations of single rWalk connections. The item param with 0.1x the influence of the supplied parameter and param_1000 with 100x the influence of the supplied parameter and 10x the value of theta.
}
\description{
Theoretical Paths using cost functions defined by Herzog 2012 with additional cost parameter
}
\details{
If there are no actual parts of a path network known, the `theoPath_herzog_param` function can be used to reconstruct pathways based on randomised shortest paths connecting known regions with higher densities of sites, e.g. monuments. An underlying cost surface is created by using the cost functions defined by I. Herzog (2012) either for walking or driving. By supplying an additional parameter (visibility, friction costs, etc.), hypotheses on the influence of different variables can be tested. This function is a useful step in the evaluation of reconstructed paths.
}
\author{
Franziska Faupel <\email{ffaupel@ufg.uni-kiel.de}>

Oliver Nakoinz <\email{oliver.nakoinz.i@gmail.com}>

Hendrik Raese <\email{h.raese@roots.uni-kiel.de}>
}
